requirements.txt
streamlit
kafka-python
streamlit-folium